|Zaidan|08/10/2023|Process master hub dan region

+++ database/migrations/2023_09_23_042310_create_subdistrict_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
     public function up()
     {
         Schema::create('subdistrict', function (Blueprint $table) {
-            $table->integer('subdistrict_id', true);
+            $table->bigInteger('subdistrict_id', true);
             $table->integer('district_id')->index('district_id');
             $table->string('name', 50);
             $table->boolean('is_active')->default(true);
